MeetMeTalkingRequestEvent.java: Use `@AsteriskMapping` instead of creating a new event class
